+++ modules/QuickEncounter.js
@@ -60,7 +60,7 @@
                 Add a config option to turn this on/off (displayEncounterXPAfterCombat) provided this is DND5E
                 Remove implicit localize in name&hint for config options
 7-Nov-2020      v0.5.5: Rename putXPInChat to computeTotalXP      
-                v0.5.5b: Tweak the dialog message          
+                v0.5.5b: Tweak the dialog          
 */